requirements.txt
aiofiles==24.1.0 ; python_version >= "3.11" and python_version < "4.0"
aiogram==3.19.0 ; python_version >= "3.11" and python_version < "4.0"
aiohappyeyeballs==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.11.16 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
apscheduler==3.11.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
magic-filter==1.0.12 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.2.0 ; python_version >= "3.11" and python_version < "4.0"
openai==1.72.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.4 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.13.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
tzlocal==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.19.0 ; python_version >= "3.11" and python_version < "4.0"
